backend-python/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import os
from dotenv import load_dotenv
from typing import List, Optional

# Load environment variables
load_dotenv()

app = FastAPI(title="Knowledge Base Search API")

# Configure CORS for frontend
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # In production, replace with actual frontend URL
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class Query(BaseModel):
    query: str

class SearchResponse(BaseModel):
    answer: str
    sources: List[str]

# Initialize clients (implement these)
# openai_client = initialize_openai()
# pinecone_client = initialize_pinecone()

@app.post("/search", response_model=SearchResponse)
async def search(query: Query):
    try:
        # TODO: Implement search logic
        # 1. Generate embeddings for the query using OpenAI
        # 2. Search Pinecone for similar chunks
        # 3. Use OpenAI to generate a response based on the retrieved chunks
        
        # Placeholder response
        return {
            "answer": "This is a placeholder response. Implement actual search logic.",
            "sources": ["Source 1", "Source 2"]
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/health")
async def health_check():
    return {"status": "healthy"} 