backend-python/services/pinecone_service.py
import pinecone
from typing import List, Dict
import os

class PineconeService:
    def __init__(self):
        # TODO: Initialize Pinecone
        # 1. Set API key and environment
        # 2. Initialize index
        # 3. Set dimension based on OpenAI embedding size
        pass
        
    async def upsert_documents(self, documents: List[Dict]):
        """
        TODO: Implement document upload
        1. Process documents into chunks
        2. Generate embeddings for chunks
        3. Upload to Pinecone with metadata
        """
        pass
        
    async def search_similar(self, query_embedding: List[float], top_k: int = 3):
        """
        TODO: Implement similarity search
        1. Use query embedding to search Pinecone
        2. Return top_k most similar chunks with metadata
        """
        pass 