backend-python/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
openai==0.27.0
pinecone-client==2.2.1
pydantic==1.8.2
python-multipart==0.0.5 